preprocess/scripts/requirements.txt
numpy
pydub
webvtt-py
numba
scipy

# Other needed deps:
# ffmpeg
# sox
# maybe others?
